Test Case Updated And Cleaned
